feat(ftux): error recovery
The primary change here is we are setting app configuration immediately
with some variables carved out for the databases being provisioned.
Then once the databases are provisioned we can replace the env vars in
the app configuration and then trigger an app deploy.

If dbs provision but app fails to deploy, we have two options to the user:
- Make source code changes and `git push` to us
- Click "redeploy" button in the UI

We are not accounting for what happens if databases fail to provision as
that is a more rare event