fix(ftux): dont derive account onboarding status based on operations
Previously we were trying always check the `Account.onboarding_status`
against the operations within an `Account` in an attempt to keep them
in-sync.  This was error prone, led to weird circular loops, and doesn't
work when we a user can have multiple `App`s within the same `Account`.

Instead, we are not going to rely on `onboarding_status` for much more
than determining if the user created the `Account` via FTUX.