feat: global `AbortController` to abort fetch and sagas
This adds an `AbortController` to our redux store so we can use
it to cancel `fetch` requests as well as any in-progress sagas.

The primary use-case is when a user logs out of we intentionally
reset the redux store it will abort any in-flight requests and prevent
old data from getting added into our store.

Example:
- We fetch all apps
- While request is in flight, user logs out
- Fetch completes and adds apps to store