fix(elevate): infinite loop (#412)
Impacted users:
- Users with OTP; and
- Users without SSH keys

This regression was caused by us trying to fix:
- User goes to app logged in via cookie
- User's token expired while still using app
- User should get sent to `/login` page

Unfortunately this caused an issue for users with OTP trying to elevate
their token.  When we try to `POST /tokens` with no otp token the API
responds with a `401` which is what we use to determine if the token is
no longer valid.

This was not an issue for a user logging in for the first time because
they are already on the `/login` page.

Current solution is to only redirect to login when a `GET` request
returns a `401`.