fix(token): dont keep token in local storage (#530)
Storing the token is wrought with edge cases.  This commit fixes the
current erroneous user flow:

- User has app and dashboard open using the same token
- In dashboard, user deletes the token (by logging out)
- In dashboard, user logs in again (creating new token)
- In app, user refreshes
- EXPECTED: user fetches `/current_token` and is allowed in app
- ACTUAL: user is redirected to login page

This erroneous behavior is caused by us trying to use the expired token
stored in local storage -- even before we hit `/current_token`.  Since
we used the wrong token the API would return a 401 and we would
immediately wipe the redux store which would then send the user to the
login page.  This would all happen before `/current_token` had a chance
to provide us with the new valid token.

Because we can never assume that the token stored in local storage is
valid, it's easier for us to not persist the token at all.