fix(db.create): look for existing db only in selected env
When a user tries to create and provision a new database we first
perform a check to see if a database with that exact name already exists
within the system.  If it does already exist, we skip creating the
database and instead move onto provision the database.

We do this because there are cases where a user can create a database
but the provision operation fails (e.g. user hits plan limits).  When
the user fixes the issue and returns to FTUX to try to
reprovision, we would then say validation failed because the database
already exists.  So if that database already exists we just try to
reprovision.