fix(environments): race condition overriding stats on detail page
The root `/accounts` endpoint does *not* include any stats besides app
and database count for performance reasons.  `/accounts/:id` does include
all the stats for an individual account.  This change creates a separate
container for stats that we set to our cache only when we fetch an
individual account endpoint.  This ensures we always have the correct
stats on an account detail page.