refactor(ftux): dont fill new dbs with existing
We were trying to be clever with adding existing databases to the new db
list and it was causing a bunch of hard-to-debug edge cases.

I believe this was to prevent the following scenario:
- User creates a database but `provision` fails (e.g. rate limit)
- User fixes rate limit and returns to ftux
- We no longer try to provision the db

For now, I think this is an edge case we can be okay with considering
the error prone solution we had lingering the codebase.