chore: log fetch errors from api (#999)
Right now we have `debugMdw` which uses a conditional logger.  Not
everyone uses the conditional logger and sometimes there's errors in
those logs that should be raised to the developer.

This change will run `console.error` anytime there is an API fetch error
so it is easier to see in our debugger.