Made small QOL improvements for diagnostics dashboards
Including:
* Impacted time range is set in user's local time, and then sent to the server in UTC
* Symptoms are free-form text, not a drop down
* The list of dashboards correctly identifies "Time Created", and adds the "Time Range"

For the final bullet point, this may not be correct anymore when we shift to a single timestamp
(time timesamp issues were noticed). However, I'm not sure whether we want to display the
single timestamp or the range identified by the diagnostic process in those scenarios, so
I'm keeping the time range for now.